requirements.txt
Flask
numpy
matplotlib
scipy
tensorflow
pypng
h5py
image